Entidad cliente funcionando

+++ src/main/java/com/comercio/app/services/impl/ClienteService.java
@@ -2,19 +2,21 @@ package com.comercio.app.services.impl;
 
 import com.comercio.app.entities.Cliente;
 import com.comercio.app.repositories.IBaseRepository;
+import com.comercio.app.repositories.impl.ClienteRepository;
 import com.comercio.app.services.IBaseService;
 import jakarta.persistence.EntityManager;
 
 import java.util.List;
+import java.util.Optional;
 
 public class ClienteService implements IBaseService<Cliente> {
 
     private EntityManager em;
     private IBaseRepository<Cliente> clienteRepository;
 
-    public ClienteService(EntityManager em, IBaseRepository<Cliente> clienteRepository) {
+    public ClienteService(EntityManager em) {
         this.em = em;
-        this.clienteRepository = clienteRepository;
+        this.clienteRepository = new ClienteRepository(this.em);
     }
 
     @Override
@@ -28,9 +30,13 @@ public class ClienteService implements IBaseService<Cliente> {
 
     @Override
     public Cliente findById(Integer id) {
-        // Verificación
         try{
-            return this.clienteRepository.findById(id);
+            if(verify(id)){
+                return this.clienteRepository.findById(id);
+            }else{
+                return null;
+            }
+
         }catch(Exception e){
             throw new Error(e.getMessage());
         }
@@ -40,43 +46,66 @@ public class ClienteService implements IBaseService<Cliente> {
     @Override
     public Cliente insert(Cliente cliente) {
         try{
+
+            this.em.getTransaction().begin();
             this.clienteRepository.insert(cliente);
+            this.em.getTransaction().commit();
+
+            return cliente;
+
         }catch(Exception e){
             this.em.getTransaction().rollback();
             throw new Error(e.getMessage());
         }
 
-        //Ver retorno del tipo Cliente
-        return null;
     }
 
     @Override
     public Cliente update(Integer id, Cliente cliente) {
+
         try{
-            this.clienteRepository.update(cliente);
+            if(verify(id)) {
+                this.em.getTransaction().begin();
+                this.clienteRepository.update(cliente);
+                this.em.getTransaction().commit();
+                return cliente;
+            }else{
+                return null;
+            }
         }catch(Exception e){
             this.em.getTransaction().rollback();
             throw new Error(e.getMessage());
         }
 
-        //Ver retorno del tipo Cliente
-        return null;
     }
+
 
     @Override
     public boolean delete(Integer id) {
-        // Verificación
-        Cliente cliente = findById(id);
 
         try{
-            this.clienteRepository.delete(cliente);
+                if(verify(id)){
+                    Cliente cliente = findById(id);
+                    this.em.getTransaction().begin();
+                    this.clienteRepository.delete(cliente);
+                    this.em.getTransaction().commit();
+                    return true;
+                }else{
+                    return false;
+                }
+
         }catch(Exception e){
             em.getTransaction().rollback();
             throw new Error(e.getMessage());
         }
 
-        //Ver retorno del tipo Cliente
-        return false;
+    }
+
+    public Boolean verify(Integer id){
+
+        Optional<Cliente> clienteVerificado = Optional.ofNullable(this.em.find(Cliente.class,id));
+
+        return clienteVerificado.isPresent();
     }
